pyproject.toml
[project]
name = "coin-trading"
version = "0.1.0"
description = "Batch-oriented Bithumb spot LLM trading system with paper trading."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler",
    "feedparser",
    "google-genai",
    "httpx",
    "litellm",
    "numpy",
    "openai",
    "pandas",
    "plotly",
    "psycopg[binary]",
    "pyjwt",
    "pydantic",
    "pydantic-settings",
    "python-dotenv",
    "sqlalchemy",
    "streamlit",
    "tenacity",
    "typer",
    "websockets",
    "langgraph>=1.1.10",
    "yfinance",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "respx",
]

[project.scripts]
coin-trading = "coin_trading.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/coin_trading"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "respx>=0.23.1",
]
